Move app message actions to their own file + add tests
