+++ src/editor/actions/app-message-actions.js
@@ -1,0 +1,16 @@
+import { addAppMessageToConsoleHistory } from "./console-message-actions";
+
+export function updateAppMessages(messageObj) {
+  return dispatch => {
+    const { message } = messageObj;
+    let { messageType, when } = messageObj;
+    if (when === undefined) when = new Date().toString();
+    if (messageType === undefined) messageType = message;
+    // add to eval history.
+    dispatch(addAppMessageToConsoleHistory(messageType));
+    dispatch({
+      type: "UPDATE_APP_MESSAGES",
+      message: { message, messageType, when }
+    });
+  };
+}
